// product-plan/sections/homepage/types.ts
export interface ComparisonCategory {
  name: string
  traditional: string
  circleup: string
}

export interface RoscaComparison {
  categories: ComparisonCategory[]
}